LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/align-items-baseline-column-vert-rl-items-expected.txt

two
lines
hello

FAIL #target > div 1 assert_equals:
<div data-offset-x="161">two<br>lines</div>
offsetLeft expected 161 but got 21
FAIL #target > div 2 assert_equals:
<div data-offset-x="201">hello</div>
offsetLeft expected 201 but got 61
